small diff

+++ src/modules/read/readability.hpp
@@ -5,7 +5,7 @@
 
 namespace Readability
 {
-double judge(cv::Mat src, int num, int flag);
+int judge(cv::Mat src, int num, int flag);
 std::pair<double, cv::Mat> edgeDetection(cv::Mat img);
 std::pair<double, cv::Mat> pointerDetection(cv::Mat src, cv::Mat origin);
 double read(cv::Mat src);
